conf: add datahangar configuration module
This commit adds a stack configuration file which captures the
set of data pipelines that configured in the system.

This commit defines two basic concepts:

1. Data pipelines: defined by filling it all fields under
   pipeline.<name>.<attr>
2. Data profiles: are the a collection of predefined templates,
   for 4 different use-cases:
   - ISP pull
   - ISP push
   - IP Carrier node to node
   - IP Carrier port to port
   One can also use "net.raw.<pmacctd_key>" to add extra fields
   to the pipeline provided that mappings exist under
   conf/datafields-db-type.json.

The conf/ package also includes some shell and python utilities to
parse the conf file and generate on the fly configuration for
other components, which can be added via configmaps to the
containers.

This is a prereq to have other components adapt its configuration
to them.